Add Docker dev environment and fix Kotlin/Ktor compilation errors
- Add Dockerfile (multi-stage: gradle:8.10.2-jdk21 builder + eclipse-temurin:21-jre-alpine runtime)
- Add docker-compose.yml with postgres + api services, healthcheck, and named volume
- Add .dockerignore to keep build context lean
- Add tools/scripts/test.sh to run gradle test via Docker with socket mount for Testcontainers
- Update tools/scripts/run-api.sh to delegate to docker compose up --build
- Copy infra/db/migrations into runtime image so Flyway can find SQL files
- Add .env to .gitignore; docker-compose.yml reads JWT_SECRET from env with dev fallback

Fix Exposed 0.54 API breakage:
- Replace deprecated Table.select{} with selectAll().where{} across all repositories
- Add explicit SqlExpressionBuilder.eq/less imports for deleteWhere lambdas
- Add missing selectAll import to UserRepository

Fix Ktor 2.3 API breakage:
- Remove CallLogging plugin (ktor-server-call-logging-jvm not reaching compile classpath)
- Convert all cookies.append(maxAge: Int) to Long

Fix Gradle settings:
- Remove redundant versionCatalogs block; Gradle 8 auto-discovers gradle/libs.versions.toml

Add dev seed migration (V202602200005):
- tenant: Demo Tenant, user: admin@example.com / password, role: ADMIN

+++ apps/api/src/main/kotlin/api/routes/AuthRoutes.kt
@@ -169,7 +169,7 @@ fun Route.authRoutes(
                     value = newToken,
                     httpOnly = true,
                     secure = isProduction,
-                    maxAge = 8 * 3600,
+                    maxAge = 8L * 3600,
                     path = "/"
                 )
 
@@ -206,7 +206,7 @@ private fun setAuthCookies(call: ApplicationCall, accessToken: String, refreshTo
         value = accessToken,
         httpOnly = true,
         secure = isProduction,
-        maxAge = 8 * 3600,
+        maxAge = 8L * 3600,
         path = "/"
     )
     call.response.cookies.append(
@@ -214,7 +214,7 @@ private fun setAuthCookies(call: ApplicationCall, accessToken: String, refreshTo
         value = refreshToken,
         httpOnly = true,
         secure = isProduction,
-        maxAge = REFRESH_TOKEN_DAYS.toInt() * 86400,
+        maxAge = REFRESH_TOKEN_DAYS * 86400,
         path = "/"
     )
 }
@@ -225,7 +225,7 @@ private fun clearAuthCookies(call: ApplicationCall, isProduction: Boolean) {
         value = "",
         httpOnly = true,
         secure = isProduction,
-        maxAge = 0,
+        maxAge = 0L,
         path = "/"
     )
     call.response.cookies.append(
@@ -233,7 +233,7 @@ private fun clearAuthCookies(call: ApplicationCall, isProduction: Boolean) {
         value = "",
         httpOnly = true,
         secure = isProduction,
-        maxAge = 0,
+        maxAge = 0L,
         path = "/"
     )
 }
